Start pulling things out of the main() function (#1)
This PR starts abstracting out bits of functionality (loading users,
generating keys) into their own functions, mostly for good practice but
also theoretically so that they can be tested (in the future).

Specifically, this PR does the following:
- Introduces a `flag` package that we can use to take in parameters,
which is needed for productionization
- Persists our hash + block keys to files so that cookies can be valid
across server runs
- Pulls out loading users into a separate function
- Allows the address to be configurable, also mainly for a server
deployment

We use the `namsral/flag` package instead of the standard library's
`flag` package so that we can pass in configuration via environment
variables and config files for extra flexibility.